% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockGroups}
\alias{blockGroups}
\title{blockGroups}
\format{An sf polygon object with 18 variables
\describe{
  \item{bgNumber}{}
  \item{areaLand}{}
  \item{areaWater}{}
  \item{year}{}

  }}
\source{
American Community Survey
}
\usage{
blockGroups
}
\description{

}
\keyword{datasets}
