% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places}
\alias{places}
\title{places}
\format{An sf polygon object with 29 variables
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  }}
\source{
American Community Survey
}
\usage{
places
}
\description{

}
\keyword{datasets}
