% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracts}
\alias{tracts}
\title{tracts}
\format{An sf polygon object with 31 variables
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  }}
\source{
American Community Survey
}
\usage{
tracts
}
\description{

}
\keyword{datasets}
